+++ src/modes/cascade.js
@@ -1,17 +1,15 @@
 /// <reference path='../types.js' />
 import { typingInterval } from '../utils'
 
+// Unify the function below with the one on cascade mode
 /** @type {TypewriterEffectFn} */
 const typewriterEffect = async ({ currentNode, text }, options) => {
-	currentNode.textContent = ''
-	currentNode.classList.add('typing')
-	for (const letter of text) {
-		currentNode.textContent += letter
+	for (let index = 0; index <= text.length; index++) {
+		const char = text[index]
+		char === '<' && (index = text.indexOf('>', index))
+		currentNode.innerHTML = text.slice(0, index)
 		await typingInterval(options.interval)
 	}
-	currentNode.nextSibling !== null && currentNode.classList.length == 1
-		? currentNode.removeAttribute('class')
-		: currentNode.classList.remove('typing')
 }
 
 /** @type {TypewriterModeFn} */